Added states and wall detection
Added “WALLAVOIDANCE” as a state for the rover in the EObstacles.m
file. Also, the rover can successfully distinguish a wall from an
obstacle. Must test with rover approaching at different angles. The new
function in “ObstacleDetection” that was added is called “IsWall.” All
these changes are in the model file called “DetectWall.slx”.