updated model
Reason for overshooting turn most likely because of the latency in
reading the sensors. The rover will first try to go around what it
thinks is a single obstacle, but instead a wall. As soon as it realizes
that it is a wall, it starts a wall avoidance maneuver. Since the rover
is at a certain angle with the wall, it still turns about 90 degrees.
This causes it to be off course after the turn. This can be solved by
forcing a delay so all sensors can acquire values or better determining
the angle to turn if this issue happens.